backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
requests==2.28.2
python-dotenv==1.0.0
flask-sqlalchemy==3.0.3
marked==4.3.0
pymdown-extensions==9.10